add missing listener methods (new architecture)
Methods `addListener` and `removeListeners` were missing from the
codegen spec. Without them, events emitted from native code do not get
sent to JS, and warnings are logged complaining about the missing
methods.

Fixes: https://github.com/RonRadtke/react-native-blob-util/issues/283
Possibly also fixes:
  https://github.com/RonRadtke/react-native-blob-util/issues/274
  https://github.com/RonRadtke/react-native-blob-util/issues/292

+++ android/src/newarch/java/com/ReactNativeBlobUtil/ReactNativeBlobUtil.java
@@ -34,7 +34,7 @@ public class ReactNativeBlobUtil extends NativeBlobUtilsSpec {
     }
 
     @ReactMethod
-    public void removeListeners(Integer count) {
+    public void removeListeners(double count) {
 
     }
